Added endpoints file to hold all endpoints. Added method in main to handle all endpoints. Added new handlers file to house the handler methods
